Add RoomContext
